+++ src/CleanRegex/Internal/Prepared/Parser/TemplateParser.php
@@ -24,8 +24,8 @@ class TemplateParser implements Parser
     public function parse(string $delimiter, QuotableFactory $quotableFactory): Quotable
     {
         TrailingBackslash::throwIfHas($this->pattern);
-        $pattern = \preg_replace_callback('/&/', function (): string {
-            return $this->nextToken()->formatAsQuotable()->quote('/');
+        $pattern = \preg_replace_callback('/&/', function () use ($delimiter): string {
+            return $this->nextToken()->formatAsQuotable()->quote($delimiter);
         }, $this->pattern);
         return new RawQuotable($pattern);
     }